FIX: Lazily Loading images did not work with spoilers
This is the result of using SVG for spoilered images, which is only
necessary for IE11.

While it's a nice hack, it's a bunch of extra logic that needn't be
applied in most browsers, and was causing issues with our image lazy
loading code, so I've elected to remove it.

Now, in IE11 the images will be not shown until clicked rather than
blurred, which I think is a decent compromise for a 5+ year old web
browser.