FIX: Remove spoilers from topic excerpts
Spoilers weren't even being blurred in topic previews that included them
so this change just removes spoilers entirely from excerpts because they
probably never should have been included in the first place:

> We have some code to preserve spoilers but I don’t know why we did
> that and it seems like a bad idea

https://meta.discourse.org/t/-/113825/5

I opted to remove spoilers inside of this plugin instead of core because
the logic should be inside of the plugin where the feature was
implemented to keep core clean of things it doesn't need to know about
if the plugin isn't installed.

This also means that iOS notifications will no longer include
`[spoiler]` as it is completely stripped out like images and links.